+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning].
 - `get_tickets` method
 - `get_transactions` method
 - `open_ticket` method
+- `send_email` method
 
 ### Changed
 - Removed `offset` param for paginated calls, use `limitstart` instead.